requirements.txt
# FastAPI và server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
xgboost==2.1.4

# OpenTelemetry Core
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# OpenTelemetry Exporters
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1

# Prometheus client
prometheus-client==0.19.0

# HTTP client (dependency cho một số OpenTelemetry components)
requests==2.31.0

# Logging và utilities
pydantic==2.5.0